fix types

+++ src/main/scala/com/github/njustus/stockdashboard/config/WatchedStocks.scala
@@ -1,7 +1,7 @@
 package com.github.njustus.stockdashboard.config
 
 case class WatchedStock(isin: String,
-                        quantity: Long,
+                        quantity: BigDecimal,
                         alias: Option[String])
 
 case class WatchedStocks(stocks: List[WatchedStock])
